now using only f32, and putting signal dimension in its type

+++ src/algebra/add.rs
@@ -16,12 +16,12 @@ impl<N> Add<Signal<N>> for Signal<N>
     }
 }
 
-impl<N> Add<f64> for Signal<N>
-    where N: Add<f64, Output = N> 
+impl<N> Add<f32> for Signal<N>
+    where N: Add<f32, Output = N> 
 {
     type Output = Signal<N>;
 
-    fn add(self, rhs: f64) -> Self::Output {
+    fn add(self, rhs: f32) -> Self::Output {
         Signal {
             value: self.value + rhs,
             domain: (self.domain.start()+rhs)..=(self.domain.end()+rhs)
@@ -29,9 +29,9 @@ impl<N> Add<f64> for Signal<N>
     }
 }
 
-impl<N, __> Add<Signal<N>> for f64
-    where f64: Add<N, Output = N>,
-    f64: Add<__>,
+impl<N, __> Add<Signal<N>> for f32
+    where f32: Add<N, Output = N>,
+    f32: Add<__>,
     N: IsSame<This = __>,
 {
     type Output = Signal<N>;
