+++ src/algebra/norm.rs
@@ -2,7 +2,7 @@ use std::ops::Div;
 use crate::Signal;
 
 
-impl<N: Div<f64, Output = N>> Signal<N> {
+impl<N: Div<f32, Output = N>> Signal<N> {
     pub fn normalize(self) -> Self {
         Signal { 
             value: self.value/(*self.domain.end()),
